python/solana-message-signing/pyproject.toml
[project]
name = "solana-message-signing"
version = "0.1.0"
description = "Sign personal messages using a Fordefi Solana Vault"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ecdsa>=0.19.0",
    "pathlib>=1.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
